resume2.0-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.5
PyMuPDF==1.23.8
pdfminer.six==20221105
python-docx==1.1.0
nltk==3.8.1
textstat==0.7.3
reportlab==4.0.7
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2
scikit-learn==1.3.2
spacy==3.7.2
pandas==2.1.3
numpy==1.25.2